Automatic workflow step registration (#253)
This PR introduces automatic creation of workflowsteps (and workflows) in B-Fabric when running applications.
To enable the functionality, one has to create the corresponding key `workflow_template_step_id` in the `bfabric` section in the app spec yaml.

The workflowstep will be created, when it does not exist, at dispatch time.